refactor: placed all hooks in a special folder
