+++ src/components/LoginForm/LoginForm.tsx
@@ -11,7 +11,8 @@ import { InputField } from '../InputField/InputField'
 import { useNavigate } from 'react-router-dom'
 import { ISignupUser, useUserSignupMutation, useUserTokenMutation } from '../../slices/music-player.api'
 import { useCookies } from 'react-cookie'
-import { useAppDispatch, useLogout } from '../../hooks'
+import { useAppDispatch } from '../../hooks/appHooks'
+import { useLogout } from '../../hooks/userHooks'
 import { setToken } from '../../slices/tokenSlice'
 import { getErrorMessage, getPasswordErrorMessage, isPasswordsIdentical, isPasswordValid, isUsernameValid } from './validators'
 import { ILoginFormState } from '../../models'
