+++ src/components/Track/Track.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react'
 
 import { ITrack } from '../../models'
 import { secondsToMinSec } from '../../utils'
-import { useFavoriteTrack } from '../Tracks/hooks'
-import { useAppDispatch } from '../../hooks'
+import { useFavoriteTrack } from '../../hooks/trackHooks'
+import { useAppDispatch } from '../../hooks/appHooks'
 import { setActiveTrackId } from '../../slices/trackSlice'
 import { HeartIcon, TrackIcon } from '../../icons'
 import { cnTrack } from './Track.classname'
