+++ src/pages/CollectionPage/CollectionPage.tsx
@@ -1,5 +1,5 @@
 import { useParams } from "react-router-dom";
-import { useCollection } from "../../components/Tracks/hooks"
+import { useCollection } from "../../hooks/trackHooks"
 import { Tracks } from "../../components/Tracks/Tracks"
       
 export const Collection = () => {
