+++ src/pages/PlaylistPage/PlaylistPage.tsx
@@ -1,4 +1,4 @@
-import { useFavoriteTracks } from "../../components/Tracks/hooks"
+import { useFavoriteTracks } from "../../hooks/trackHooks"
 import { Tracks } from "../../components/Tracks/Tracks"
       
 export const PlaylistPage = () => <Tracks title="Мой плейлист" tracksHook={useFavoriteTracks}/>
