+++ src/pages/TracksPage/TracksPage.tsx
@@ -1,4 +1,4 @@
-import { useFilteredTracks } from '../../components/Tracks/hooks'
+import { useFilteredTracks } from '../../hooks/trackHooks'
 import { Tracks } from '../../components/Tracks/Tracks'
     
 export const TracksPage = () => <Tracks title="Треки" showFilter tracksHook={useFilteredTracks}/>
